requirements.txt
torch
numpy
pandas
scikit-learn
librosa
soundfile
matplotlib
seaborn
tqdm
pyyaml
optuna
streamlit
plotly
scipy
pytest
requests
